+++ src/WebformBuilder.unit.js
@@ -3,11 +3,14 @@ import Harness from '../test/harness';
 import WebformBuilder from './WebformBuilder';
 import Builders from './builders';
 import { Formio } from './Formio';
-import { uniqueApiKeys, uniqueApiKeysLayout, uniqueApiKeysSameLevel, columnsForm, resourceKeyCamelCase } from '../test/formtest';
+import { uniqueApiKeys, uniqueApiKeysLayout, uniqueApiKeysSameLevel, columnsForm, resourceKeyCamelCase, uniqueApiKeysTranslation } from '../test/formtest';
 import sameApiKeysLayoutComps from '../test/forms/sameApiKeysLayoutComps';
 import testApiKeysUniquifying from '../test/forms/testApiKeysUniquifying';
 import formBasedOnWizard from '../test/forms/formBasedOnWizard';
 import formWithFormController from '../test/forms/formWithFormController';
+
+global.requestAnimationFrame = (cb) => cb();
+global.cancelAnimationFrame = () => {};
 
 describe('WebformBuilder tests', function() {
   this.retries(3);
@@ -31,6 +34,27 @@ describe('WebformBuilder tests', function() {
       assert.equal(builder.webform.components[0].disabled, true);
       done();
     }, 500);
+  });
+  it('Should show API Key is not unique: {{key}} error when api keys are the same', (done) => {
+    const builder = Harness.getBuilder();
+    builder.i18next.currentLanguage = { apiKey: 'translated api key error {{key}}' };
+    builder.webform.setForm(uniqueApiKeysTranslation).then(()=>{
+      builder.highlightInvalidComponents();
+      const component = builder.webform.getComponent(['textField']);
+      assert.equal(component.visibleErrors.length, 1);
+      done();
+    }).catch(done);
+  });
+
+  it('Should show translated api key error {{key}} when apiKey is overridden in i18next translations', (done) => {
+    const builder = Harness.getBuilder();
+    builder.i18next.currentLanguage = { apiKey: 'translated api key error {{key}}' };
+    builder.webform.setForm(uniqueApiKeysTranslation).then(() => {
+      builder.highlightInvalidComponents();
+      const component = builder.webform.getComponent(['textField']);
+      assert.equal(component.visibleErrors[0].message,'translated api key error textField');
+      done();
+    }).catch(done);
   });
 
   it('Should not show unique API error when components with same keys are inside and outside of the Data component', (done) => {
@@ -283,6 +307,9 @@ describe('Select Component selectData property', () => {
         }, {
           label: 'Label 2',
           value: 'value2',
+        }, {
+          label: 'Label 3',
+          value: 'value3',
         }];
 
         resolve(values);
@@ -374,6 +401,45 @@ describe('Select Component selectData property', () => {
 
           setTimeout(() => {
             assert.equal(builder.editForm.data.selectData, undefined);
+            Harness.saveComponent();
+            setTimeout(() => {
+              done();
+            }, 150);
+          }, 250);
+        }, 250);
+      }, 150);
+    }).catch(done);
+  });
+
+  it('Should calculate multiple selectData property for url dataSource', (done) => {
+    const builder = Harness.getBuilder();
+    builder.setForm({}).then(() => {
+      Harness.buildComponent('select');
+
+      setTimeout(() => {
+        const multiple = builder.editForm.getComponent('multiple');
+        multiple.setValue(true);
+        const dataSrc = builder.editForm.getComponent('dataSrc');
+        dataSrc.setValue('url');
+        const url = builder.editForm.getComponent(['data.url']);
+        const valueProperty = builder.editForm.getComponent('valueProperty');
+        url.setValue('htts//fakeurl.com');
+        valueProperty.setValue('value');
+
+        setTimeout(() => {
+          const defaultValue = builder.editForm.getComponent('defaultValue');
+          defaultValue.setValue(['value1', 'value3']);
+          defaultValue.updateItems(null, true);
+
+          setTimeout(() => {
+            assert.deepEqual(builder.editForm.data.selectData, {
+              value1: {
+                label: 'Label 1',
+              },
+              value3: {
+                label: 'Label 3',
+              },
+            });
             Harness.saveComponent();
             setTimeout(() => {
               done();