+++ src/components/select/Select.unit.js
@@ -32,6 +32,8 @@ import {
   comp20,
   comp21,
   comp22,
+  comp23,
+  comp24,
 } from './fixtures';
 
 // eslint-disable-next-line max-statements
@@ -1042,6 +1044,63 @@ describe('Select Component', () => {
     }).catch(done);
   });
 
+  it('Should provide correct metadata.selectData for multiple Select with default value', (done) => {
+    const form = _.cloneDeep(comp23);
+    const element = document.createElement('div');
+
+    Formio.createForm(element, form).then(form => {
+      const submit = form.getComponent('submit');
+      const clickEvent = new Event('click');
+      const submitBtn = submit.refs.button;
+      submitBtn.dispatchEvent(clickEvent);
+
+      setTimeout(()=> {
+        const metadata = form.submission.metadata.selectData.select;
+        assert.deepEqual(metadata, {
+          value1: {
+            label: 'Label 1',
+          },
+          value3: {
+            label: 'Label 3',
+          },
+        });
+        done();
+      }, 200);
+    }).catch(done);
+  });
+
+  it('Should set correct label from metadata for multiple Select with default value', (done) => {
+    const form = _.cloneDeep(comp23);
+    const element = document.createElement('div');
+
+    Formio.createForm(element, form).then(form => {
+      const select = form.getComponent('select');
+      form.submission = {
+        data: {
+          select: ['value1', 'value2'],
+        },
+        metadata: {
+          selectData: {
+            select: {
+              value1: {
+                label: 'Label 1',
+              },
+              value2: {
+                label: 'Label 2',
+              },
+            },
+          },
+        },
+      };
+
+      setTimeout(()=> {
+        assert.equal(select.templateData['value1'].label, 'Label 1');
+        assert.equal(select.templateData['value2'].label, 'Label 2');
+        done();
+      }, 200);
+    }).catch(done);
+  });
+
   it('OnBlur validation should work properly with Select component', function(done) {
     this.timeout(0);
     const element = document.createElement('div');
@@ -1276,5 +1335,68 @@ describe('Select Component with Entire Object Value Property', () => {
       done();
     });
   });
-});
 
+  it('Should render label for Select components when Data Source is Resource in read only mode', (done) => {
+    const element = document.createElement('div');
+    Formio.createForm(element, comp24, { readOnly: true }).then((form) => {
+      const select = form.getComponent('select');
+      form.setSubmission({
+        metadata: {
+          selectData: {
+            select: {
+              data: {
+                textField1: 'A',
+              },
+            },
+          },
+          timezone: 'Europe/Kiev',
+          offset: 180,
+          origin: 'http://localhost:3001',
+          referrer: '',
+          browserName: 'Netscape',
+          userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/125.0.0.0 Safari/537.36',
+          pathName: '/',
+          onLine: true,
+          headers: {
+            host: 'qvecgdgwpwujbpi.localhost:3000',
+            connection: 'keep-alive',
+            'content-length': '457',
+            'sec-ch-ua': '"Google Chrome";v="125", "Chromium";v="125", "Not.A/Brand";v="24"',
+            accept: 'application/json',
+            'content-type': 'application/json',
+            'sec-ch-ua-mobile': '?0',
+            'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/125.0.0.0 Safari/537.36',
+            'sec-ch-ua-platform': '"Windows"',
+            origin: 'http://localhost:3001',
+            'sec-fetch-site': 'cross-site',
+            'sec-fetch-mode': 'cors',
+            'sec-fetch-dest': 'empty',
+            referer: 'http://localhost:3001/',
+            'accept-encoding': 'gzip, deflate, br, zstd',
+            'accept-language': 'en-US,en;q=0.9,ru-RU;q=0.8,ru;q=0.7',
+          },
+        },
+        data: {
+          select: 1,
+          select1: {
+            textField1: 'A',
+            textField2: '1',
+            submit: true,
+          },
+          submit: true,
+        },
+        state: 'submitted',
+      });
+
+      setTimeout(() => {
+        const previewSelect = select.element.querySelector('[aria-selected="true"] span');
+
+        assert.equal(previewSelect.innerHTML, 'A', 'Should show label as a selected value' +
+          ' for Select component');
+
+        done();
+      }, 300);
+    })
+      .catch((err) => done(err));
+  });
+});